+++ engine/math/Matrix3.cs
@@ -4,49 +4,49 @@ namespace engine.math
 {
     public class Matrix3
     {
-        private float[,] m;
+        private double[,] m;
 
-        public Matrix3() : this(new float[3, 3]) {}
+        public Matrix3() : this(new double[3, 3]) {}
 
-        public Matrix3(float[,] m)
+        public Matrix3(double[,] m)
         {
             this.m = m;
         }
 
         public Matrix3 initIdentity()
         {
-            return new Matrix3(new float [,] { {1, 0, 0},
+            return new Matrix3(new double [,] { {1, 0, 0},
                                                {0, 1, 0},
                                                {0, 0, 1} });
         }
 
         public Matrix3 initTranslation(Vector2 translation)
         {
-            return new Matrix3(new float [,] { {1, 0, translation.getX()},
+            return new Matrix3(new double [,] { {1, 0, translation.getX()},
                                                {0, 1, translation.getY()},
                                                {0, 0, 1} });
         }
 
-        public Matrix3 initRotation(float rotation)
+        public Matrix3 initRotation(double rotation)
         {
-            float sin = (float)Math.Sin(rotation);
-            float cos = (float)Math.Cos(rotation);
+            double sin = (double)Math.Sin(rotation);
+            double cos = (double)Math.Cos(rotation);
 
-            return new Matrix3(new float [,] { {cos, -sin, 0},
+            return new Matrix3(new double [,] { {cos, -sin, 0},
                                                {sin, cos, 0},
                                                {0, 0, 1} });
         }
 
         public Matrix3 initScale(Vector2 scale)
         {
-            return new Matrix3(new float [,] { {scale.getX(), 0, 0},
+            return new Matrix3(new double [,] { {scale.getX(), 0, 0},
                                                {0, scale.getY(), 0},
                                                {0, 0, 1} });
         }
 
         public Vector2 mul(Vector2 r)
         {
-            float[] v = new float[3];
+            double[] v = new double[3];
 
             for(int x = 0; x < 3; x++)
             {
@@ -75,12 +75,12 @@ namespace engine.math
             return result;
         }
 
-        public void set(float value, int x, int y)
+        public void set(double value, int x, int y)
         {
             m[x, y] = value;
         }
 
-        public float get(int x, int y)
+        public double get(int x, int y)
         {
             return m[x, y];
         }
